+++ lib/src/core/managers/full_screen_dialog_manager/full_screen_dialog_show_params.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_easy_dialogs/flutter_easy_dialogs.dart';
-import 'package:flutter_easy_dialogs/src/core/managers/easy_dialog_manager_base.dart';
 import 'package:flutter_easy_dialogs/src/core/managers/full_screen_dialog_manager/full_screen_dialog_manager.dart';
 
 /// Show params for [FullScreenDialogManager]