+++ lib/src/core/managers/positioned_dialog_manager/positioned_dialog_show_params.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_easy_dialogs/flutter_easy_dialogs.dart';
-import 'package:flutter_easy_dialogs/src/core/managers/easy_dialog_manager_base.dart';
 import 'package:flutter_easy_dialogs/src/core/managers/positioned_dialog_manager/positioned_dialog_manager.dart';
 
 /// Show params of [PositionedDialogManager]