+++ lib/src/core/overlay/easy_overlay.dart
@@ -13,62 +13,9 @@ import 'package:flutter_easy_dialogs/src/core/managers/positioned_dialog_manager
 import 'package:flutter_easy_dialogs/src/core/overlay/overlay.dart';
 import 'package:flutter_easy_dialogs/src/util/position_to_animation_converter/position_to_animation_converter.dart';
 
-// Service - helper for easy use different custom dialogs
-class FlutterEasyDialogs extends StatelessWidget {
-  /// Theme of [FlutterEasyDialogs]
-  final FlutterEasyDialogsThemeData? theme;
-
-  final CustomManagerBuilder? customManagerBuilder;
-
-  /// Child widget
-  final Widget child;
-
-  /// Creates an instance of [FlutterEasyDialogs]
-  const FlutterEasyDialogs({
-    required this.child,
-    this.customManagerBuilder,
-    this.theme,
-    super.key,
-  });
-
-  static final _key = GlobalKey<_EasyOverlayState>();
-
-  /// Gets [EasyDialogsController]
-  static EasyDialogsController get dialogsController =>
-      _key.currentState!._easyDialogsController;
-
-  /// For using in [MaterialApp.builder]
-  static const builder = _builder;
-
-  static TransitionBuilder _builder({
-    FlutterEasyDialogsThemeData? theme,
-    CustomManagerBuilder? customManagerBuilder,
-  }) {
-    return (context, child) => FlutterEasyDialogs(
-          theme: theme,
-          customManagerBuilder: customManagerBuilder,
-          child: child ?? const SizedBox.shrink(),
-        );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return FlutterEasyDialogsTheme(
-      data: theme ?? FlutterEasyDialogsThemeData.basic(),
-      child: Material(
-        child: EasyOverlay(
-          key: _key,
-          customManagersBuilder: customManagerBuilder,
-          initialEntries: [
-            EasyOverlayAppEntry(
-              builder: (context) => child,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
+part 'easy_overlay_insert_strategy.dart';
+part 'easy_overlay_remove_strategy.dart';
+part 'keys.dart';
 
 /// Function for providing custom agents
 typedef CustomManagerBuilder = List<EasyDialogManagerBase> Function(
@@ -89,36 +36,32 @@ class EasyOverlay extends Overlay {
   });
 
   @override
-  OverlayState createState() => _EasyOverlayState();
+  EasyOverlayState createState() => EasyOverlayState();
 }
 
-class _EasyOverlayState extends OverlayState implements IEasyOverlayController {
-  final _currentPositionedEntries = <EasyDialogPosition, OverlayEntry>{};
-  final _currentCustomEntries = <OverlayEntry>[];
+class EasyOverlayState extends OverlayState implements IEasyOverlayController {
+  @visibleForTesting
+  final currentEntries = <Object?, Object?>{};
 
-  OverlayEntry? _currentFullScreenEntry;
   OverlayEntry? _appEntry;
 
-  late final EasyDialogsController _easyDialogsController;
+  @override
+  EasyOverlay get widget => super.widget as EasyOverlay;
+
+  late final EasyDialogsController easyDialogsController;
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties
       ..add(
-        DiagnosticsProperty<Map<EasyDialogPosition, OverlayEntry>>(
-          'currentPositionedEntries',
-          _currentPositionedEntries,
+        DiagnosticsProperty<Map<Object?, Object?>>(
+          '_currentEntries',
+          currentEntries,
         ),
       )
       ..add(
         DiagnosticsProperty<OverlayEntry?>('appEntry', _appEntry),
-      )
-      ..add(
-        DiagnosticsProperty<OverlayEntry?>(
-          'currentFullScreenEntry',
-          _currentFullScreenEntry,
-        ),
       );
   }
 
@@ -131,52 +74,9 @@ class _EasyOverlayState extends OverlayState implements IEasyOverlayController {
   @override
   void didChangeDependencies() {
     final theme = FlutterEasyDialogsTheme.of(context);
-    _easyDialogsController.updateTheme(theme);
+    easyDialogsController.updateTheme(theme);
 
     super.didChangeDependencies();
-  }
-
-  @override
-  void insertPositionedDialog({
-    required Widget dialog,
-    required EasyDialogPosition position,
-  }) {
-    assert(
-      !_currentPositionedEntries.containsKey(position),
-      'only single one $EasyOverlayEntry with the same $EasyDialogPosition can be presented at the same time',
-    );
-
-    final entry = EasyOverlayEntry(
-      builder: (_) => dialog,
-    );
-
-    _currentPositionedEntries[position] = entry;
-
-    insert(entry);
-  }
-
-  @override
-  void insertFullScreenDialog({
-    required Widget dialog,
-  }) {
-    assert(
-      _currentFullScreenEntry == null,
-      'only single one full screen $EasyOverlayEntry can be presented',
-    );
-
-    final entry = EasyOverlayEntry(
-      builder: (_) => dialog,
-    );
-
-    insert(entry);
-
-    _currentFullScreenEntry = entry;
-  }
-
-  @override
-  void removeFullScreenDialog() {
-    _currentFullScreenEntry!.remove();
-    _currentFullScreenEntry = null;
   }
 
   @override
@@ -197,19 +97,6 @@ class _EasyOverlayState extends OverlayState implements IEasyOverlayController {
     super.insert(entry, below: below, above: above);
   }
 
-  @override
-  void removePositionedDialog({
-    required EasyDialogPosition position,
-  }) {
-    if (_currentPositionedEntries.entries.isEmpty) return;
-
-    final entry = _currentPositionedEntries.remove(position);
-
-    if (entry == null || !entry.mounted) return;
-
-    entry.remove();
-  }
-
   void _init() {
     final positionToAnimationConverter = PositionToAnimationConverter();
     final positionedAnimationFactory =
@@ -218,7 +105,7 @@ class _EasyOverlayState extends OverlayState implements IEasyOverlayController {
     final positionedDismissibleFactory = PositionedDismissibleFactory();
     final modalBannerFactory = EasyModalBannerFactory();
 
-    final bannerAgent = PositionedDialogManager(
+    final bannerManager = PositionedDialogManager(
       overlayController: this,
       dialogFactory: EasyBannerFactory(
         animationFactory: positionedAnimationFactory,
@@ -226,12 +113,11 @@ class _EasyOverlayState extends OverlayState implements IEasyOverlayController {
       ),
     );
 
-    final modalBannerAgent = FullScreenDialogManager(
+    final modalBannerManager = FullScreenDialogManager(
       overlayController: this,
       dialogFactory: modalBannerFactory,
     );
-    final customManagersRaw =
-        (widget as EasyOverlay).customManagersBuilder?.call(this);
+    final customManagersRaw = widget.customManagersBuilder?.call(this);
 
     final customManagers = <EasyDialogManagerBase>[];
 
@@ -247,9 +133,9 @@ class _EasyOverlayState extends OverlayState implements IEasyOverlayController {
       }
     }
 
-    _easyDialogsController = EasyDialogsController(
-      bannerManager: bannerAgent,
-      modalBannerManager: modalBannerAgent,
+    easyDialogsController = EasyDialogsController(
+      bannerManager: bannerManager,
+      modalBannerManager: modalBannerManager,
       customManagers: Map.fromIterable(
         customManagers,
         key: (customAgent) => customAgent.runtimeType,
@@ -258,55 +144,15 @@ class _EasyOverlayState extends OverlayState implements IEasyOverlayController {
   }
 
   @override
-  int insertCustomDialog(Widget dialog) {
-    final entry = OverlayEntry(builder: (_) => dialog);
-    _currentCustomEntries.add(entry);
-    insert(entry);
-
-    return _currentCustomEntries.indexOf(entry);
-  }
+  void insertDialog(EasyOverlayInsertStrategy strategy) => strategy.apply(this);
 
   @override
-  void removeCustomDialog(int id) {
-    final entry = ((id < _currentCustomEntries.length)
-        ? _currentCustomEntries[id]
-        : null);
-
-    if (entry == null) return;
-
-    _currentCustomEntries.removeAt(id);
-    entry.remove();
-  }
+  void removeDialog(EasyOverlayRemoveStrategy strategy) => strategy.apply(this);
 }
 
 /// Interface for manipulating overlay with dialogs
 abstract class IEasyOverlayController extends TickerProvider {
-  /// Insert positioned dialog into overlay
-  ///
-  /// Only single one dialog of concrete position can persists at the same time
-  void insertPositionedDialog({
-    required Widget dialog,
-    required EasyDialogPosition position,
-  });
+  void insertDialog(EasyOverlayInsertStrategy strategy);
 
-  /// Remove positioned dialog from overlay
-  void removePositionedDialog({
-    required EasyDialogPosition position,
-  });
-
-  /// Insert full screen dialog into overlay
-  void insertFullScreenDialog({
-    required Widget dialog,
-  });
-
-  /// Remove full screen dialog into overlay
-  void removeFullScreenDialog();
-
-  /// Insert custom dialog
-  ///
-  /// returns id of inserted [dialog]
-  int insertCustomDialog(Widget dialog);
-
-  /// Remove custom dialog using [id]
-  void removeCustomDialog(int id);
+  void removeDialog(EasyOverlayRemoveStrategy strategy);
 }