+++ test/src/core/agents/signle_auto_disposal_controller_agent_test.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_easy_dialogs/src/core/managers/mixin/single_auto_disposal_controller_mixin.dart';
+
+void main() {
+  late SingleAutoDisposalControllerMixin sut;
+}