+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-## 1.0.9
+## 1.1.0
 
 * Fixed typos
 * Refactored project structure
 * `Agents` renamed to `Managers`
 * Added `BlockAndroidBackButtonMixin` and `SingleAutoDisposalControllerMixin`
+* Reworked inserting/removing dialogs in overlay
+* Provided tests on insert/remove strategies
 
 ## 1.0.8